+++ sentry-ruby/spec/sentry/transport/http_transport_spec.rb
@@ -1,17 +1,20 @@
 require 'spec_helper'
 
 RSpec.describe Sentry::HTTPTransport do
-  let(:client) { Sentry::Client.new(Sentry.configuration) }
-  let(:event) { client.event_from_message("test") }
-  subject { described_class.new(Sentry.configuration) }
+  let(:configuration) do
+    Sentry::Configuration.new.tap do |config|
+      config.dsn = 'http://12345@sentry.localdomain/sentry/42'
+    end
+  end
+  let(:client) { Sentry::Client.new(configuration) }
+  let(:event) { client.event_from_message("foobarbaz") }
+  let(:data) do
+    subject.encode(event.to_hash)
+  end
+
+  subject { described_class.new(configuration) }
 
   describe "customizations" do
-    before do
-      Sentry.init do |c|
-        c.dsn = 'http://12345@sentry.localdomain/sentry/42'
-      end
-    end
-
     it 'sets a custom User-Agent' do
       expect(subject.conn.headers[:user_agent]).to eq("sentry-ruby/#{Sentry::VERSION}")
     end
@@ -19,7 +22,7 @@ RSpec.describe Sentry::HTTPTransport do
     it 'allows to customise faraday' do
       builder = spy('faraday_builder')
       expect(Faraday).to receive(:new).and_yield(builder)
-      Sentry.configuration.transport.faraday_builder = proc { |b| b.request :instrumentation }
+      configuration.transport.faraday_builder = proc { |b| b.request :instrumentation }
 
       subject
 
@@ -27,14 +30,63 @@ RSpec.describe Sentry::HTTPTransport do
     end
   end
 
-  describe "failed request handling" do
-    before do
-      Sentry.init do |c|
-        c.dsn = 'http://12345@sentry.localdomain/sentry/42'
-        c.transport.http_adapter = [:test, stubs]
-        c.transport.transport_class = described_class
+  describe "request payload" do
+    let(:compressed_stubs) do
+      Faraday::Adapter::Test::Stubs.new do |stub|
+        stub.post('sentry/api/42/envelope/') do |env|
+          expect(env.request_headers["Content-Type"]).to eq("application/x-sentry-envelope")
+          expect(env.request_headers["Content-Encoding"]).to eq("gzip")
+
+          envelope = Zlib.gunzip(env.body)
+          expect(envelope).to include(event.event_id)
+          expect(envelope).to include("foobarbaz")
+        end
       end
     end
+
+    let(:uncompressed_stubs) do
+      Faraday::Adapter::Test::Stubs.new do |stub|
+        stub.post('sentry/api/42/envelope/') do |env|
+          expect(env.request_headers["Content-Type"]).to eq("application/x-sentry-envelope")
+          expect(env.request_headers["Content-Encoding"]).to eq("")
+
+          envelope = env.body
+          expect(envelope).to include(event.event_id)
+          expect(envelope).to include("foobarbaz")
+        end
+      end
+    end
+
+    it "compresses data by default" do
+      configuration.transport.http_adapter = [:test, compressed_stubs]
+
+      subject.send_data(data)
+      compressed_stubs.verify_stubbed_calls
+    end
+
+    it "doesn't compress small event" do
+      configuration.transport.http_adapter = [:test, uncompressed_stubs]
+
+      event.instance_variable_set(:@threads, nil) # shrink event
+
+      subject.send_data(data)
+      uncompressed_stubs.verify_stubbed_calls
+    end
+
+    it "doesn't compress data if the encoding is not gzip" do
+      configuration.transport.http_adapter = [:test, uncompressed_stubs]
+      configuration.transport.encoding = "json"
+
+      subject.send_data(data)
+      uncompressed_stubs.verify_stubbed_calls
+    end
+  end
+
+  describe "failed request handling" do
+    before do
+      configuration.transport.http_adapter = [:test, stubs]
+    end
+
     context "receive 4xx responses" do
       let(:stubs) do
         Faraday::Adapter::Test::Stubs.new do |stub|
@@ -43,7 +95,7 @@ RSpec.describe Sentry::HTTPTransport do
       end
 
       it 'raises an error' do
-        expect { subject.send_data(event.to_hash) }.to raise_error(Sentry::Error, /the server responded with status 404/)
+        expect { subject.send_data(data) }.to raise_error(Sentry::Error, /the server responded with status 404/)
 
         stubs.verify_stubbed_calls
       end
@@ -57,7 +109,7 @@ RSpec.describe Sentry::HTTPTransport do
       end
 
       it 'raises an error' do
-        expect { subject.send_data(event.to_hash) }.to raise_error(Sentry::Error, /the server responded with status 500/)
+        expect { subject.send_data(data) }.to raise_error(Sentry::Error, /the server responded with status 500/)
 
         stubs.verify_stubbed_calls
       end
@@ -71,7 +123,7 @@ RSpec.describe Sentry::HTTPTransport do
       end
 
       it 'raises an error with header' do
-        expect { subject.send_data(event.to_hash) }.to raise_error(Sentry::Error, /error_in_header/)
+        expect { subject.send_data(data) }.to raise_error(Sentry::Error, /error_in_header/)
 
         stubs.verify_stubbed_calls
       end