+++ sentry-ruby/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Refactor tracing implementation [#1309](https://github.com/getsentry/sentry-ruby/pull/1309)
 - Allow configuring BreadcrumbBuffer's size limit [#1310](https://github.com/getsentry/sentry-ruby/pull/1310)
 - Return nil from logger methods instead of breadcrumb buffer [#1299](https://github.com/getsentry/sentry-ruby/pull/1299)
+- Compress event payload by default [#1314](https://github.com/getsentry/sentry-ruby/pull/1314)
 
 ## 4.2.2